requirements.txt
numpy
tqdm
underthesea
rank-bm25
openai
python-dotenv
scikit-learn
psycopg2-binary
pgvector